+++ fj-core/src/test/java/test/org/fugerit/java/core/util/mvn/TestMavenProps.java
@@ -1,26 +1,37 @@
 package test.org.fugerit.java.core.util.mvn;
 
+import org.fugerit.java.core.util.mvn.FJCoreMaven;
 import org.fugerit.java.core.util.mvn.MavenProps;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import lombok.extern.slf4j.Slf4j;
 
+import java.util.Optional;
+
 @Slf4j
 class TestMavenProps {
 
 	@Test
 	void testMavenOk() {
-		String prop = MavenProps.getPropery( "org.fugerit.java", "fj-core", "artifactId" );
-		log.info( "version {}", prop );
-		Assertions.assertEquals( "fj-core" , prop );
+		Optional prop = MavenProps.getPropertyOptional(FJCoreMaven.FJ_CORE_GROUP_ID, FJCoreMaven.FJ_CORE_ARTIFACT_ID, MavenProps.ARTIFACT_ID );
+		log.info( "version ok {}", prop.get() );
+		Assertions.assertEquals( "fj-core" , prop.get() );
 	}
 	
 	@Test
 	void testMavenKo() {
-		String prop = MavenProps.getPropery( "org.fugerit.java.no.exists", "fj-core", "artifactId" );
-		log.info( "version {}", prop );
+		// use deprecated method for coverage
+		String prop = MavenProps.getPropery( "org.fugerit.java.no.exists", FJCoreMaven.FJ_CORE_ARTIFACT_ID, MavenProps.ARTIFACT_ID );
+		log.info( "version ko {}", prop );
 		Assertions.assertNull( prop );
+	}
+
+	@Test
+	void testFjCoreVersion() {
+		Optional fjCoreVersion = FJCoreMaven.getFJCoreVersion();
+		Assertions.assertTrue( fjCoreVersion.isPresent() );
+		log.info( "fjCoreVersion ok {}", fjCoreVersion.get() );
 	}
 	
 }